Layout to pay slip and fix

+++ app/src/main/java/br/com/if91wallet/vo/TransactionRequest.kt
@@ -5,10 +5,10 @@ import com.google.gson.annotations.SerializedName
 class TransactionRequest (
     @SerializedName("card_number") val cardNumber: Long,
     val cvv: Int,
-    val value: Double,
+    val value: Float,
     @SerializedName("expiry_date") val expirationDate: String,
     @SerializedName("destination_user_id") val destinationUserId: Long
 ) {
-    constructor(value: Double, cardVo: CardVo, userVo: UserVo)
+    constructor(value: Float, cardVo: CardVo, userVo: UserVo)
             : this(cardVo.cardNumber, cardVo.cvv, value, cardVo.expirationDate, userVo.id)
 }
